requirements.txt
h5py==2.10.0
imageio==2.5.0
Keras==2.2.4
matplotlib==3.1.0
numpy==1.16.4
pandas==0.25.0
scikit-image==0.15.0
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.9.0
tensorflow-gpu==1.13.1